+++ Documentation/readme.md
@@ -5,6 +5,10 @@ Below you can find a list of all the available cmdlets. Many commands provide bu
 Get-Help Connect-SPOnline -Detailed
 ```
 
+##
+Cmdlet|Description
+:-----|:----------
+**[Connect&#8209;PnPMicrosoftGraph](ConnectPnPMicrosoftGraph.md)** |Uses the Microsoft Authentication Library (Preview) to connect to Azure AD and to get an OAuth 2.0 Access Token to consume the Microsoft Graph API
 ##Apps
 Cmdlet|Description
 :-----|:----------
@@ -124,7 +128,6 @@ Cmdlet|Description
 ##Provisioning
 Cmdlet|Description
 :-----|:----------
-**[Connect&#8209;PnPMicrosoftGraph](ConnectPnPMicrosoftGraph.md)** |Uses the Microsoft Authentication Library (Preview) to connect to Azure AD and to get an OAuth 2.0 Access Token to consume the Microsoft Graph API
 **[Remove&#8209;SPOFileFromProvisioningTemplate](RemoveSPOFileFromProvisioningTemplate.md)** |Removes a file from an in-memory PnP Provisioning Template
 **[Add&#8209;SPOFileToProvisioningTemplate](AddSPOFileToProvisioningTemplate.md)** |Adds a file to an in-memory PnP Provisioning Template
 **[Convert&#8209;SPOFolderToProvisioningTemplate](ConvertSPOFolderToProvisioningTemplate.md)** |Creates a pnp package file of an existing template xml, and includes all files in the current folder
